Update overview.md
Removed link to LoB definition on Wikipedia
Reformulated the framework information.  Needs verification.
Moved the Learn how to develop to the end
Removed the section on CQRS (this is described later and is not critical to the definition here)
added .io to instances of Dolittle
Make some small changes to the EDA and Microservices sections.  More work is needed here.